lol_stats/collector/models.py
from django.db import models


class Summoners(models.Model):

    summoner_id = models.CharField(max_length=250)
    account_id = models.CharField(max_length=250)
    puuid = models.CharField(max_length=250)
    name = models.CharField(max_length=250)
    profile_icon_id = models.CharField(max_length=250)
    revision_date = models.DateTimeField()
    summoner_level = models.IntegerField()


    def get_summoner(self,name):
        import requests
        headers = {
            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/111.0.0.0 Safari/537.36",
            "Accept-Language": "pt-BR,pt;q=0.9,en-US;q=0.8,en;q=0.7",
            "Accept-Charset": "application/x-www-form-urlencoded; charset=UTF-8",
            "Origin": "https://developer.riotgames.com",
            "X-Riot-Token": KEY
        }