requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
Markdown==3.4.3
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
django-environasgiref==3.6.0
colorama==0.4.6
Django==4.1.7
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
Markdown==3.4.3
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
